move status to component
