+++ pages/build.js
@@ -9,11 +9,9 @@ import Diff from "../components/Diff";
 import CommitData from "../components/CommitData";
 import theme from "../styles/theme";
 import api from "../util/api";
-import Pending from "../components/svg/Pending";
-import Failed from "../components/svg/Failed";
-import Approved from "../components/svg/Approved";
 import Images from "../components/svg/Images";
 import BreadCrumbs from "../components/Breadcrumbs";
+import BuildStatus from "../components/BuildStatus";
 
 class BuildContainer extends Component {
   constructor(props) {
@@ -83,11 +81,7 @@ class Build extends Component {
           /
         </BreadCrumbs>
         <BuildHeader>
-          <BuildStatus status={status}>
-            {(status == "passed" || status == "approved") && <Approved />}
-            {status == "pending" && <Pending />}
-            {status == "failed" && <Failed />}
-          </BuildStatus>
+          <BuildStatus status={status} />
           <div>
             <PullRequestTitle>
               {pr_branch || branch || "Unknown Build"}
@@ -120,13 +114,6 @@ class Build extends Component {
 export default withRouter(BuildContainer);
 export { Build };
 
-let getColor = status => {
-  if (status == "pending") return theme.yellow;
-  if (status == "passed") return theme.green;
-  if (status == "failed") return theme.red;
-  return theme.gray8;
-};
-
 let BuildHeader = styled("div")`
   background: ${theme.gray3};
   padding: 2em;
@@ -146,19 +133,6 @@ let BuildNumber = styled("span")`
   font-weight: bold;
   font-size: 0.9em;
   margin-right: 0.5em;
-`;
-
-let BuildStatus = styled("div")`
-  background: ${theme.gray9};
-  border: 3px solid #fff;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 3.5em;
-  width: 3.5em;
-  border-radius: 3.5em;
-  color: ${p => getColor(p.status)};
-  margin-right: 0.66em;
 `;
 
 let ApproveButton = styled("div")`
